Split architecture-specific stuff into its own header
